backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-storage>=2.16.0
python-multipart
